### 2020-06-08 #58 @whedon
[REVIEW]: Circuitscape in Julia: High Performance Connectivity Modelling to Support Conservation Decisions
**Submitting author:** @ranjanan (<a href="http://orcid.org/0000-0002-4409-3937">Ranjan Anantharaman</a>)
**Repository:** <a href="https://github.com/Circuitscape/CircuitscapeJuliaPaper" target ="_blank">https://github.com/Circuitscape/CircuitscapeJuliaPaper</a>
