### 2020-06-08 #58 @matbesancon
@vlandau thanks for the review and sorry for the delay, can you verify that the paper meets the criteria from the list above and check them when it applies?
