### 2020-06-08 #58 @vlandau
Hmm, when I click the link to accept the invitation I am getting a 404 error that the page doesn't exist -- I haven't been able to accept an invitation so maybe that's why I'm not seeing how to check the checklist above.
