### 2020-06-08 #58 @ranjanan
Almost all the DOIs that whedon complained are missing exist, except for the cholesky factorization citation (https://onlinelibrary.wiley.com/doi/full/10.1002/wics.18) which has a different DOI than the one whedon detected. I'll update the paper with the DOI and generate it here.
