### 2020-06-11 #58 @matbesancon
>  I'm not sure what the page limit is, so it is unchecked for now.

if I remember correctly it's 8 pages, so this paper is fine

You can verify that references are adequate, disregarding the technical DOI issue, that is, is the paper referencing appropriate references to point the reader where needed
