### 2020-06-15 #58 @ranjanan
Updated the fautly DOI and generated the paper, like you asked.
