### 2020-06-17 #58 @matbesancon
fine grained -> fine-grained

Over the past decade Circuitscape -> Over the past decade, Circuitscape

Then, in collaboration with Tanmay Mohapatra, it was translated to Python for flexible scripting, platform independence, and released under an open-source license.

in the C programmming language --> 2 Ms

is high performance -> high-performance

our user base, and works well on every platform (comma)

Julia is an open source programming language --> open-source (multiple occurrences)

are often problem specific -> problem-specific

the choleksy factorization -> Choleksy (multiple occurrences)

backsubstitution -> back-substitution

process independently \cite{pelletier2014applying} , and (no space before and)

produce fine scale and localized -> fine-scale

switch to 64-bit integer ->integers

feedback loop with policy makers and result -> policymakers, and

Its high level mathematical -> high-level

presents upto -> up to

is open source and is available -> is open-source
