### 2020-06-29 08370f3
More typos and suggestions
