+++ README.md
@@ -1,1 +1,3 @@
-# CircuitscapeJuliaPaper
+# Circuitscape Paper - JuliaCon 2020
+
+This is the JuliaCon 2020 submission for the Circuitscape.jl package. For software, and usage instructions, please refer to [the software README](https://github.com/Circuitscape/Circuitscape.jl/blob/master/README.md) and [the website](https://circuitscape.org/)
