### 2020-06-29 #58 @ranjanan
@matbesancon I made the changes, but I changed the URL from the original package repo to https://github.com/Circuitscape/CircuitscapeJuliaPaper.
