### 2020-07-26 #58 @matbesancon
if you go on Zenodo, you can link your github, and then register your project, this will allow you to register an archive
