### 2020-07-26 #58 @ranjanan
Yes, here's the release DOI: 10.5281/zenodo.3960829. Should I create a DOI for the paper as well, or will this do?
