### 2020-07-26 #58 @ranjanan
All DOIs that whedon is complaining about are valid, except for the "cholesky factorization" one. But I updated the DOI for that reference here: https://github.com/Circuitscape/CircuitscapeJuliaPaper/blob/master/paper/paper.bib#L221. Doesn't look like whedon looks at that field?
